+++ src/main/java/project/board/service/MemberService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface MemberService {
     SignUpResponseDto addMember(MemberSignUpRequestDto dto);
 
-    MemberResponseDto updateMember(int id, MemberUpdateProfileRequestDto dto);
+    MemberResponseDto updateMember(Long id, MemberUpdateProfileRequestDto dto);
 
     MemberResponseDto searchMemberById(long id);
 
-    MemberResponseDto changePassword(int id, PasswordChangeRequestDto dto);
+    MemberResponseDto changePassword(Long id, PasswordChangeRequestDto dto);
 
     List<MemberResponseDto> searchAllMember();
